2.0.2: - graph: Fixed potential buffer overflow.        - audio_mixer: Added volume graph for debugging.        - write_frame: Cosmetics.        - misc: Fixed AUTO bugs caused by arethmetic overflows.        - ffmpeg_producer: Fixed nb_frames when disabling looping.
git-svn-id: https://casparcg.svn.sourceforge.net/svnroot/casparcg/server/branches/2.0.2@1796 362d55ac-95cf-4e76-9f9a-cbaa9c17b72d

Former-commit-id: 2e34146cbc85f7069f3a6d81d95a2270de979681

+++ core/producer/separated/separated_producer.cpp
@@ -81,7 +81,7 @@ struct separated_producer : public frame_producer
 		return disable_audio(last_frame_);
 	}
 
-	virtual int64_t nb_frames() const override
+	virtual uint32_t nb_frames() const override
 	{
 		return std::min(fill_producer_->nb_frames(), key_producer_->nb_frames());
 	}
